+++ moe/moe_token_unpermute_with_ep/op_host/moe_token_unpermute_with_ep_tiling.cpp
@@ -103,7 +103,7 @@ static inline ge::graphStatus MoeTokenUnpermuteWithEpInputParamCheck(
     auto dataTensor0 = context->GetInputTensor(0);
     auto dataTensor1 = context->GetInputTensor(1);
     auto nodeName = context->GetNodeName();
-    const int* tmpTopK = context->GetAttrs()->GetAttrPointer<int>(0);
+    const int64_t* tmpTopK = context->GetAttrs()->GetAttrPointer<int64_t>(0);
     int inputTopK = static_cast<int64_t>(*tmpTopK);
 
     OP_CHECK_IF(
@@ -171,7 +171,7 @@ static inline void Init(
     param.input.isUnpermute = isUnpermute;
 
     auto attrPtr = context->GetAttrs();
-    const int* tmpTopK = attrPtr->GetAttrPointer<int>(UNPERMUTE_WITH_EP_ARRT_TOPK);
+    const int64_t* tmpTopK = attrPtr->GetAttrPointer<int64_t>(UNPERMUTE_WITH_EP_ARRT_TOPK);
     int inputTopK = static_cast<int64_t>(*tmpTopK);
     auto rangePtr = attrPtr->GetAttrPointer<gert::ContinuousVector>(UNPERMUTE_WITH_EP_ARRT_RANGE);
     if (rangePtr != nullptr) {
